oxfr impl removed and get random secondary zone from existing zones for unit test cases
